remove a bogus cast to int for file positions in debug output fix typo in doxygen comment
git-svn-id: http://svn.mythtv.org/svn/trunk@14050 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/avformatdecoder.cpp
@@ -2010,7 +2010,7 @@ void AvFormatDecoder::HandleGopStart(AVPacket *pkt)
             VERBOSE(VB_PLAYBACK|VB_TIMESTAMP, LOC + 
                     QString("positionMap[ %1 ] == %2.")
                     .arg(prevgoppos / keyframedist)
-                    .arg((int)startpos));
+                    .arg(startpos));
 
             PosMapEntry entry = {prevgoppos / keyframedist,
                                  prevgoppos, startpos};
